backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.7.1
openai==1.3.5
google-generativeai==0.3.2
pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1